requirements.txt
web3
pandas
python-dotenv
db-dtypes
google-cloud-bigquery